feat: Add latest train deep neural network
For this commit we are adding the main changes
to the agent. The agent now trains successfully, and
the main script receives command line arguments.

Additionally we added:

    - New project report as a README.md
    - Several percepts samples
    - Updated checkpoint and network files

Finally, this commit also fixes the issue seen
where the upper part of the paddle wasn't
detecting the lower portion of the ball.